✨ custom feedBrain for chat and brain management
