+++ frontend/app/brains-management/[brainId]/components/BrainManagementTabs/components/KnowledgeTab/hooks/useFeedBrain.ts
@@ -1,0 +1,94 @@
+import { UUID } from "crypto";
+import { useParams } from "next/navigation";
+import { useState } from "react";
+import { useTranslation } from "react-i18next";
+
+import {
+  FeedItemCrawlType,
+  FeedItemUploadType,
+} from "@/app/chat/[chatId]/components/ActionsBar/types";
+import { useChatApi } from "@/lib/api/chat/useChatApi";
+import { useKnowledgeToFeedInput } from "@/lib/components/KnowledgeToFeedInput/hooks/useKnowledgeToFeedInput.ts";
+import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
+import { useKnowledgeContext } from "@/lib/context/KnowledgeProvider/hooks/useKnowledgeContext";
+import { useToast } from "@/lib/hooks";
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const useFeedBrain = ({
+  dispatchHasPendingRequests,
+  closeFeedInput,
+}: {
+  dispatchHasPendingRequests?: () => void;
+  closeFeedInput?: () => void;
+}) => {
+  const { publish } = useToast();
+  const { t } = useTranslation(["upload"]);
+
+  const { currentBrainId } = useBrainContext();
+  const { setKnowledgeToFeed, knowledgeToFeed } = useKnowledgeContext();
+  const [hasPendingRequests, setHasPendingRequests] = useState(false);
+
+  const { createChat } = useChatApi();
+  const params = useParams();
+  const chatId = params?.chatId as UUID | undefined;
+
+  const { crawlWebsiteHandler, uploadFileHandler } = useKnowledgeToFeedInput();
+
+  const files: File[] = (
+    knowledgeToFeed.filter((c) => c.source === "upload") as FeedItemUploadType[]
+  ).map((c) => c.file);
+
+  const urls: string[] = (
+    knowledgeToFeed.filter((c) => c.source === "crawl") as FeedItemCrawlType[]
+  ).map((c) => c.url);
+
+  const feedBrain = async (): Promise<void> => {
+    if (currentBrainId === null) {
+      publish({
+        variant: "danger",
+        text: t("selectBrainFirst"),
+      });
+
+      return;
+    }
+
+    if (knowledgeToFeed.length === 0) {
+      publish({
+        variant: "danger",
+        text: t("addFiles"),
+      });
+
+      return;
+    }
+
+    try {
+      dispatchHasPendingRequests?.();
+      closeFeedInput?.();
+      setHasPendingRequests(true);
+      const currentChatId = chatId ?? (await createChat("New Chat")).chat_id;
+      const uploadPromises = files.map((file) =>
+        uploadFileHandler(file, currentBrainId, currentChatId)
+      );
+      const crawlPromises = urls.map((url) =>
+        crawlWebsiteHandler(url, currentBrainId, currentChatId)
+      );
+
+      await Promise.all([...uploadPromises, ...crawlPromises]);
+
+      setKnowledgeToFeed([]);
+    } catch (e) {
+      publish({
+        variant: "danger",
+        text: JSON.stringify(e),
+      });
+    } finally {
+      setHasPendingRequests(false);
+    }
+  };
+
+  return {
+    feedBrain,
+    hasPendingRequests,
+    setHasPendingRequests,
+  };
+};
