+++ frontend/lib/components/KnowledgeToFeedInput/hooks/useKnowledgeToFeedInput.ts.ts
@@ -1,23 +1,11 @@
-/* eslint-disable max-lines */
 import { UUID } from "crypto";
-import { useParams, useRouter } from "next/navigation";
-import { useCallback, useState } from "react";
+import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useChatApi } from "@/lib/api/chat/useChatApi";
 import { useCrawlApi } from "@/lib/api/crawl/useCrawlApi";
-import { useNotificationApi } from "@/lib/api/notification/useNotificationApi";
 import { useUploadApi } from "@/lib/api/upload/useUploadApi";
-import { useChatContext } from "@/lib/context";
-import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
-import { useKnowledgeContext } from "@/lib/context/KnowledgeProvider/hooks/useKnowledgeContext";
 import { getAxiosErrorParams } from "@/lib/helpers/getAxiosErrorParams";
 import { useToast } from "@/lib/hooks";
-
-import {
-  FeedItemCrawlType,
-  FeedItemUploadType,
-} from "../../../../app/chat/[chatId]/components/ActionsBar/types";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const useKnowledgeToFeedInput = () => {
@@ -100,98 +88,5 @@ export const useKnowledgeToFeedInput = () => {
   return {
     crawlWebsiteHandler,
     uploadFileHandler,
-  };
-};
-
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export const useFeedBrainInChat = ({
-  dispatchHasPendingRequests,
-  closeFeedInput,
-}: {
-  dispatchHasPendingRequests?: () => void;
-  closeFeedInput?: () => void;
-}) => {
-  const { publish } = useToast();
-  const { t } = useTranslation(["upload"]);
-  const router = useRouter();
-
-  const { currentBrainId } = useBrainContext();
-  const { setKnowledgeToFeed, knowledgeToFeed } = useKnowledgeContext();
-  const [hasPendingRequests, setHasPendingRequests] = useState(false);
-
-  const { createChat } = useChatApi();
-  const params = useParams();
-  const chatId = params?.chatId as UUID | undefined;
-
-  const { setNotifications } = useChatContext();
-  const { getChatNotifications } = useNotificationApi();
-  const fetchNotifications = async (currentChatId: UUID): Promise<void> => {
-    const fetchedNotifications = await getChatNotifications(currentChatId);
-    setNotifications(fetchedNotifications);
-  };
-
-  const { crawlWebsiteHandler, uploadFileHandler } = useKnowledgeToFeedInput();
-
-  const files: File[] = (
-    knowledgeToFeed.filter((c) => c.source === "upload") as FeedItemUploadType[]
-  ).map((c) => c.file);
-
-  const urls: string[] = (
-    knowledgeToFeed.filter((c) => c.source === "crawl") as FeedItemCrawlType[]
-  ).map((c) => c.url);
-
-  const feedBrain = async (): Promise<void> => {
-    if (currentBrainId === null) {
-      publish({
-        variant: "danger",
-        text: t("selectBrainFirst"),
-      });
-
-      return;
-    }
-
-    if (knowledgeToFeed.length === 0) {
-      publish({
-        variant: "danger",
-        text: t("addFiles"),
-      });
-
-      return;
-    }
-
-    try {
-      dispatchHasPendingRequests?.();
-      closeFeedInput?.();
-      setHasPendingRequests(true);
-      const currentChatId = chatId ?? (await createChat("New Chat")).chat_id;
-      const uploadPromises = files.map((file) =>
-        uploadFileHandler(file, currentBrainId, currentChatId)
-      );
-      const crawlPromises = urls.map((url) =>
-        crawlWebsiteHandler(url, currentBrainId, currentChatId)
-      );
-
-      await Promise.all([...uploadPromises, ...crawlPromises]);
-
-      setKnowledgeToFeed([]);
-
-      if (chatId === undefined) {
-        void router.push(`/chat/${currentChatId}`);
-      } else {
-        await fetchNotifications(currentChatId);
-      }
-    } catch (e) {
-      publish({
-        variant: "danger",
-        text: JSON.stringify(e),
-      });
-    } finally {
-      setHasPendingRequests(false);
-    }
-  };
-
-  return {
-    feedBrain,
-    hasPendingRequests,
   };
 };
